requirements.txt
Django==3.1.6
django-crispy-forms==1.11.0
pytz==2021.1
six==1.15.0
django-cors-headers==3.7.0
django-extensions==3.1.1
django-clearcache==1.1.1
Faker==6.5.0